requirements.txt
streamlit==1.26.0
pandas==2.2.2
yfinance==0.2.28
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
